remove anychart
