--- src/app/components/day-of-week-histogram-apex/day-of-week-histogram-apex.component.html
@@ -1,7 +1,0 @@
-<apx-chart
-  [series]="series"
-  [chart]="chart"
-  [plotOptions]="plotOptions"
-  [xaxis]="xaxis"
-  [colors]="colors"
-></apx-chart>
