+++ src/app/components/day-of-week-histogram/day-of-week-histogram.component.html
@@ -1,1 +1,7 @@
-<div id="chart-container" #chartContainer></div>
+<apx-chart
+  [series]="series"
+  [chart]="chart"
+  [plotOptions]="plotOptions"
+  [xaxis]="xaxis"
+  [colors]="colors"
+></apx-chart>
