--- src/app/components/day-of-week-histogram/day-of-week-histogram.component.scss
@@ -1,8 +1,0 @@
-#chart-container
-{
-  margin-top: 1.5em;
-  margin-left: auto;
-  margin-right: auto;
-  width: 95%;
-  height: 85%;
-}
