+++ src/app/models/analytics-item-repository.service.ts
@@ -4,7 +4,6 @@ import { VolatileModelRepositoryService } from './model-repository.service';
 import { AnalyticsItem } from './analytics-item';
 
 import { DayOfWeekHistogramComponent } from '../components/day-of-week-histogram/day-of-week-histogram.component';
-import { DayOfWeekHistogramApexComponent } from '../components/day-of-week-histogram-apex/day-of-week-histogram-apex.component';
 import { HourOfDayRadarComponent } from '../components/hour-of-day-radar/hour-of-day-radar.component';
 
 @Injectable({
@@ -13,26 +12,15 @@ import { HourOfDayRadarComponent } from '../components/hour-of-day-radar/hour-of
 export class AnalyticsItemRepositoryService extends VolatileModelRepositoryService<AnalyticsItem> {
 
   public async loadAll() {
+
     this.save(
       new AnalyticsItem(
         DayOfWeekHistogramComponent,
         {
           name: 'Day of Week Histogram',
-          description: 'Anychart Demo',
+          description: '',
           icon: 'calendar',
           url: 'day-of-week-histogram',
-        }
-      )
-    );
-
-    this.save(
-      new AnalyticsItem(
-        DayOfWeekHistogramApexComponent,
-        {
-          name: 'Day of Week Histogram',
-          description: 'Apex Demo',
-          icon: 'calendar',
-          url: 'day-of-week-histogram-apex',
         }
       )
     );
@@ -42,7 +30,7 @@ export class AnalyticsItemRepositoryService extends VolatileModelRepositoryServi
         HourOfDayRadarComponent,
         {
           name: 'Hour of Day Radar',
-          description: 'Apex Demo',
+          description: '',
           icon: 'calendar',
           url: 'hour-of-day-radar',
         }
